fix(rules): correct multi-line tests for body-max-line-length
Previously the multi-line test fixtures in the test suite for
the `body-max-line-length` rule were defined but never used;
instead, the single line fixtures were being used in tests
supposed to test multi-line commit messages.

This was presumably an innocent copy-and-paste error, so fix the
multi-line tests to use the multi-line fixtures.