feat(rules): make body-max-line-length ignore lines with URLs
Whilst the `body-max-line-length` is almost always desirable, there
is one fairly frequent scenario in which it gets in the way: when
lines contain URLs exceeding the maximum line length.

In this case, the URL cannot be split across lines without breaking
it, and minifying it results in obfuscation and a dependency on a
third-party service, both of which are undesirable.

So make an exception in this rule for any line which contains a URL.
Allow other content on the same line, in order to support use of
various rich-text formats such as Markdown, which are often used
in order to render links in more elegant ways.